Add back sanitization of our env variables
This partially fixes https://github.com/cdr/code-server/issues/4519.

We use an environment variable to distinguish the parent process from
the child so without sanitizing when you spawn code-server from VS Code
it always looks like we are in the parent and the process will fail
since it is not actually the child.