Remove last opened functionality
We must handle this before we get to the browser because the first thing
it does is create a workspace provider based on the provided workspace
before any services are created which means we cannot fetch settings.
So if we grab the last opened after that is done it has the wrong
workspace and this breaks the close folder functionality (and possibly
other things).

Since it needs to be done on the server I figure I will implement in
code-server to minimize the patch.