Bypass spurious reload button
This reload button does nothing which I think must mean there is some
logic somewhere that is supposed to load these extensions locally that
is not working or not implemented.

Until we find it bypassing is probably a good move because otherwise it
makes it look like the extension is not running and working even though
it really is (it is just running remotely).

This will be a temporary measure to solve
https://github.com/coder/code-server/issues/4600.