fix: conflict in src/vs/platform/remote/browser/browserSocketFactory.ts
I am not 100% if the new code is similar to ours. We had to patch this to work
against the current path.

Their new implementation may do the same thing but I couldn't tell from looking
at it.

We may want to revisit this later when we switch to patches. Hopefully we can
remove this.