fix:  src/vs/platform/remote/common/remoteAgentConnection.ts
It appears they've added a new parameter to `createSocket` which takes a string
for a debugLabel.

Not sure how it's used or made available but it's interesting because it takes
the `connectionTypeToString` which we're adding to the query.

Makes me think we might be able to remove this patch if our backend can get
access to the debugLabel somehow.